requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-migrate>=4.0.5
flask-socketio>=5.3.6
werkzeug>=3.0.0
python-dotenv>=1.0.0

flask
gunicorn
psycopg2-binary
requests